refactor(arrow): #1028 — extract the upload-guard tests per the layout convention
`make preflight`'s lint-tests-layout gate refuses an inline `mod tests { … }` block;
the guard's tests move to `upload_guard/tests.rs`. Caught by preflight, which I had
read through a pipe — `$?` was `tail`'s, not `make`'s, so the first push went out on a
branch that did not pass. Same trap this program has hit before.

## Verification

✅ `make preflight` exit 0, checked directly rather than through a pipe
✅ upload_guard 5/5
